requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
pycodestyle==2.10.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.4.0